@@ -9,12 +9,19 @@
     "build": "vitepress build content",
     "serve": "vitepress serve content --port 5012"
   },
+  "type": "module",
   "devDependencies": {
+    "@tailwindcss/typography": "^0.5.7",
+    "@types/node": "^18.7.19",
+    "autoprefixer": "^10.4.12",
     "gray-matter": "^4.0.3",
-    "sass": "^1.49.9",
-    "typescript": "^4.6.2",
-    "vitepress": "^0.22.3",
-    "vue": "^3.2.31"
+    "postcss": "^8.4.16",
+    "sass": "^1.55.0",
+    "tailwindcss": "^3.1.8",
+    "terser": "^5.15.0",
+    "typescript": "^4.8.3",
+    "vitepress": "1.0.0-alpha.15",
+    "vue": "^3.2.39"
   },
   "engines": {
     "node": ">=12.0.0"
