@@ -1,17 +1,16 @@
-import '../theme/styles/main.scss';
+import './styles.scss';
+import './code.scss';
 
 import { Theme } from 'vitepress';
 import Layout from './Layout.vue';
 import NotFound from './NotFound.vue';
-import About from './components/About.vue';
-import Projects from './components/Projects.vue';
-import Tags from './components/Tags.vue';
+import Projects from './pages/Projects.vue';
+import Tags from './pages/Tags.vue';
 
 const theme: Theme = {
   Layout,
   NotFound,
   enhanceApp: ({ app }) => {
-    app.component('About', About);
     app.component('Projects', Projects);
     app.component('Tags', Tags);
   },
